Revision 2 of MonsterOS
MonsterOS shell prompt can now partially execute external commands. All
external commands are files stored in a file table on the hard disk;
that are loaded in memory at OS startup. If a external command is a
program it will be executed and the program will be loaded at 0x9999.
Else the program is treated as a text file and its contents are
displayed. So far MonsterOS does nothing when the file in the file table
is not a program. Amongst other things this revision contains some minor
tweaks with some of the software interrupts and.